+++ backend/database/recurringTaskHistory.go
@@ -16,7 +16,8 @@ type RecurringTaskHistory struct {
 	DoneAt          time.Time `json:"doneAt"`
 	CreatedAt       time.Time `json:"createdAt"`
 	UpdatedAt       time.Time `json:"updatedAt"`
-	RecurringTaskId uuid.UUID `json:"userId"`
+	UserId          uuid.UUID `json:"userId"`
+	RecurringTaskId uuid.UUID `json:"recurringTaskId"`
 }
 
 func (t *RecurringTaskHistory) merge(s *RecurringTaskHistory) {
@@ -25,11 +26,11 @@ func (t *RecurringTaskHistory) merge(s *RecurringTaskHistory) {
 	}
 }
 
-func (s *DBService) GetRecurringTaskHistoryById(id uuid.UUID, pid uuid.UUID) (*RecurringTaskHistory, error) {
+func (s *DBService) GetRecurringTaskHistoryById(id uuid.UUID, uid uuid.UUID) (*RecurringTaskHistory, error) {
 	res, err := s.db.Query(
-		"SELECT * FROM recurring_tasks_history WHERE id=$1 AND recurring_task_id=$2",
+		"SELECT * FROM recurring_tasks_history WHERE id=$1 AND user_id=$2",
 		id,
-		pid,
+		uid,
 	)
 	if err != nil {
 		return nil, err
@@ -44,17 +45,18 @@ func (s *DBService) GetRecurringTaskHistoryById(id uuid.UUID, pid uuid.UUID) (*R
 	var doneAt time.Time
 	var createdAt time.Time
 	var updatedAt time.Time
+	var userId uuid.UUID
 	var recurringTaskId uuid.UUID
 
-	if err := res.Scan(&tId, &desc, &done, &doneAt, &createdAt, &updatedAt, &recurringTaskId); err != nil {
+	if err := res.Scan(&tId, &desc, &done, &doneAt, &createdAt, &updatedAt, &userId, &recurringTaskId); err != nil {
 		return nil, err
 	}
-	task := RecurringTaskHistory{Id: tId, Description: desc, Done: done, DoneAt: doneAt, CreatedAt: createdAt, UpdatedAt: updatedAt, RecurringTaskId: recurringTaskId}
+	task := RecurringTaskHistory{Id: tId, Description: desc, Done: done, DoneAt: doneAt, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId, RecurringTaskId: recurringTaskId}
 	return &task, nil
 }
 
-func (s *DBService) GetRecurringTaskHistory(id uuid.UUID) ([]RecurringTaskHistory, error) {
-	res, err := s.db.Query("SELECT * FROM recurring_tasks_history WHERE recurring_task_id=$1", id)
+func (s *DBService) GetRecurringTaskHistoriesByUser(id uuid.UUID) ([]RecurringTaskHistory, error) {
+	res, err := s.db.Query("SELECT * FROM recurring_tasks_history WHERE user_id=$1", id)
 	if err != nil {
 		return nil, err
 	}
@@ -70,12 +72,46 @@ func (s *DBService) GetRecurringTaskHistory(id uuid.UUID) ([]RecurringTaskHistor
 		var doneAt time.Time
 		var createdAt time.Time
 		var updatedAt time.Time
+		var userId uuid.UUID
 		var recurringTaskId uuid.UUID
 
-		if err := res.Scan(&tId, &desc, &done, &doneAt, &createdAt, &updatedAt, &recurringTaskId); err != nil {
+		if err := res.Scan(&tId, &desc, &done, &doneAt, &createdAt, &updatedAt, &userId, &recurringTaskId); err != nil {
 			return nil, err
 		}
-		task := RecurringTaskHistory{Id: tId, Description: desc, Done: done, DoneAt: doneAt, CreatedAt: createdAt, UpdatedAt: updatedAt, RecurringTaskId: recurringTaskId}
+		task := RecurringTaskHistory{Id: tId, Description: desc, Done: done, DoneAt: doneAt, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId, RecurringTaskId: recurringTaskId}
+		tasks = append(tasks, task)
+	}
+	return tasks, nil
+}
+
+func (s *DBService) GetRecurringTasksHistory(id uuid.UUID, uid uuid.UUID) ([]RecurringTaskHistory, error) {
+	res, err := s.db.Query(
+		"SELECT * FROM recurring_tasks_history WHERE recurring_task_id=$1 AND user_id=$2",
+		id,
+		uid,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	defer res.Close()
+
+	tasks := []RecurringTaskHistory{}
+
+	for res.Next() {
+		var tId uuid.UUID
+		var desc string
+		var done bool
+		var doneAt time.Time
+		var createdAt time.Time
+		var updatedAt time.Time
+		var userId uuid.UUID
+		var recurringTaskId uuid.UUID
+
+		if err := res.Scan(&tId, &desc, &done, &doneAt, &createdAt, &updatedAt, &userId, &recurringTaskId); err != nil {
+			return nil, err
+		}
+		task := RecurringTaskHistory{Id: tId, Description: desc, Done: done, DoneAt: doneAt, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId, RecurringTaskId: recurringTaskId}
 		tasks = append(tasks, task)
 	}
 	return tasks, nil
@@ -86,9 +122,10 @@ func (s *DBService) InsertRecurringTaskHistory(task RecurringTaskHistory) (uuid.
 	err := crdb.ExecuteTx(context.Background(), s.db, nil,
 		func(tx *sql.Tx) error {
 			err := tx.QueryRow(
-				"INSERT INTO recurring_tasks_history (description, done, done_at, recurring_task_id) VALUES ($1, $2, now(), $3) RETURNING id",
+				"INSERT INTO recurring_tasks_history (description, done, user_id, recurring_task_id) VALUES ($1, $2, $3, $4) RETURNING id",
 				task.Description,
 				task.Done,
+				task.UserId,
 				task.RecurringTaskId,
 			).Scan(&id)
 
@@ -101,7 +138,7 @@ func (s *DBService) InsertRecurringTaskHistory(task RecurringTaskHistory) (uuid.
 }
 
 func (s *DBService) UpdateRecurringTaskHistory(reqTask RecurringTaskHistory) error {
-	task, err := s.GetRecurringTaskHistoryById(reqTask.Id, reqTask.RecurringTaskId)
+	task, err := s.GetRecurringTaskHistoryById(reqTask.Id, reqTask.UserId)
 	if err != nil {
 		return err
 	}
@@ -109,10 +146,11 @@ func (s *DBService) UpdateRecurringTaskHistory(reqTask RecurringTaskHistory) err
 	return crdb.ExecuteTx(context.Background(), s.db, nil,
 		func(tx *sql.Tx) error {
 			_, err := tx.Exec(
-				"UPDATE recurring_tasks_history SET description = $1, updated_at=now() WHERE id = $2 AND recurring_task_id=$3",
+				"UPDATE recurring_tasks_history SET description = $1, updated_at=now() WHERE id = $2 AND recurring_task_id=$3 AND user_id=$4",
 				task.Description,
 				task.Id,
 				task.RecurringTaskId,
+				task.UserId,
 			)
 			return err
 		})
@@ -122,20 +160,21 @@ func (s *DBService) UpdateRecurringTaskHistoryDone(task RecurringTaskHistory) er
 	return crdb.ExecuteTx(context.Background(), s.db, nil,
 		func(tx *sql.Tx) error {
 			_, err := tx.Exec(
-				"UPDATE recurring_tasks_history SET done = $1, done_at=now(), updated_at=now() WHERE id = $2 AND recurring_task_id=$3",
+				"UPDATE recurring_tasks_history SET done = $1, done_at=now(), updated_at=now() WHERE id = $2 AND recurring_task_id=$3 AND user_id=$4",
 				task.Done,
 				task.Id,
 				task.RecurringTaskId,
+				task.UserId,
 			)
 			return err
 		})
 }
 
-func (s *DBService) DeleteRecurringTaskHistory(id uuid.UUID, pid uuid.UUID) error {
+func (s *DBService) DeleteRecurringTaskHistory(id uuid.UUID, uid uuid.UUID) error {
 	res, err := s.db.Query(
-		"DELETE FROM recurring_tasks_history WHERE id = $1 AND recurring_task_id=$2",
+		"DELETE FROM recurring_tasks_history WHERE id = $1 AND user_id=$2",
 		id,
-		pid,
+		uid,
 	)
 	if err != nil {
 		return err
@@ -144,10 +183,11 @@ func (s *DBService) DeleteRecurringTaskHistory(id uuid.UUID, pid uuid.UUID) erro
 	return nil
 }
 
-func (s *DBService) DeleteCompleteRecurringTaskHistory(id uuid.UUID) error {
+func (s *DBService) DeleteCompleteRecurringTaskHistory(id uuid.UUID, uid uuid.UUID) error {
 	res, err := s.db.Query(
-		"DELETE FROM recurring_tasks_history WHERE recurring_task_id=$1",
+		"DELETE FROM recurring_tasks_history WHERE recurring_task_id=$1 AND user_id=$2",
 		id,
+		uid,
 	)
 	if err != nil {
 		return err
